Addressed some PR comments and fixed typos.

+++ sst/test/src/tg2_cache_access.cc
@@ -17,9 +17,9 @@ TEST_CASE(TG2, "cache_access_of_load_to_same_address", "src",
     # Execute first load
     ldr x1, [x0]
 
-    # Multiple instructions to wait for store to execute.
+    # Multiple instructions to wait for load to execute.
     # Can't be replaced by loop to ensure that caching was
-    # established by store
+    # established by load.
     mov x4, #0
     add x4, x4, #1
     add x4, x4, #1
@@ -88,7 +88,7 @@ TEST_CASE(TG2, "load_after_store_on_same_address_should_be_return_from_cache",
 
     # Multiple instructions to wait for store to execute.
     # Can't be replaced by loop to ensure that caching was
-    # established by store
+    # established by store.
     mov x4, #0
     add x4, x4, #1
     add x4, x4, #1
@@ -133,8 +133,8 @@ TEST_CASE(TG2, "load_after_store_on_same_address_should_be_return_from_cache",
   uint64_t lastLoadData = reads[2]->data_;
 
   EXPECT_LT(lastLoadCycleLatency, firstLoadCycleLatency);
-  EXPECT_EQ(1024, firstLoadData);
-  EXPECT_EQ(1024, lastLoadData);
+  EXPECT_EQ((uint64_t)1024, firstLoadData);
+  EXPECT_EQ((uint64_t)1024, lastLoadData);
 }
 
 TEST_CASE(TG2,
@@ -265,7 +265,7 @@ TEST_CASE(TG2,
   // First request to load value from heap, Check if it is equal to initialised
   // data.
   uint64_t firstLoadData = reads[1]->data_;
-  EXPECT_EQ(1024, firstLoadData);
+  EXPECT_EQ((uint64_t)1024, firstLoadData);
 
   // The access latency specified in fastL1WithParams_config.py is 2 clock
   // cycle. By the time this data is aggregated and delivered to SimEng it
@@ -276,26 +276,26 @@ TEST_CASE(TG2,
   // load at addr: [x2]
   uint64_t secondLoadCycleLatency = reads[2]->endCycle_ - reads[2]->startCycle_;
   uint64_t secondLoadData = reads[2]->data_;
-  EXPECT_LTE(secondLoadCycleLatency, 3);
-  EXPECT_EQ(1024, secondLoadData);
+  EXPECT_LTE(secondLoadCycleLatency, (uint64_t)3);
+  EXPECT_EQ((uint64_t)1024, secondLoadData);
   // load at addr: [x2, #8]
   uint64_t thirdLoadCycleLatency = reads[3]->endCycle_ - reads[3]->startCycle_;
   uint64_t thirdLoadData = reads[3]->data_;
-  EXPECT_LTE(thirdLoadCycleLatency, 3);
-  EXPECT_EQ(1024, thirdLoadData);
+  EXPECT_LTE(thirdLoadCycleLatency, (uint64_t)3);
+  EXPECT_EQ((uint64_t)1024, thirdLoadData);
   // load at addr: [x2, #16]
   uint64_t fourthLoadCycleLatency = reads[4]->endCycle_ - reads[4]->startCycle_;
   uint64_t fourthLoadData = reads[4]->data_;
-  EXPECT_LTE(fourthLoadCycleLatency, 3);
-  EXPECT_EQ(1024, fourthLoadData);
+  EXPECT_LTE(fourthLoadCycleLatency, (uint64_t)3);
+  EXPECT_EQ((uint64_t)1024, fourthLoadData);
   // load at addr: [x2, #24]
   uint64_t fifthLoadCycleLatency = reads[5]->endCycle_ - reads[5]->startCycle_;
   uint64_t fifthLoadData = reads[5]->data_;
-  EXPECT_LTE(fifthLoadCycleLatency, 3);
-  EXPECT_EQ(1024, fifthLoadData);
+  EXPECT_LTE(fifthLoadCycleLatency, (uint64_t)3);
+  EXPECT_EQ((uint64_t)1024, fifthLoadData);
   // load at addr: [x2, #32]
   uint64_t sixthLoadCycleLatency = reads[6]->endCycle_ - reads[6]->startCycle_;
   uint64_t sixthLoadData = reads[6]->data_;
-  EXPECT_LTE(sixthLoadCycleLatency, 3);
-  EXPECT_EQ(1024, sixthLoadData);
+  EXPECT_LTE(sixthLoadCycleLatency, (uint64_t)3);
+  EXPECT_EQ((uint64_t)1024, sixthLoadData);
 }
